Added a file with will take an object and pickle it with the current date and time.
Added a clientFromCredentials function to return a client by loading a yaml file with a path reference to GCP project credentials.json.

Added a TableManagement function which takes a valid client, presumably now from the specified project details and credentials location in our config-yaml.yaml and loaded from that, creates the schema and pushes to a target dataset and tablename. Note only the client.project is used in the create_bigquery_table call, as we want to retain the ability to over-ride these fields (there will only ever be on project per client, but there may be many dataset_ids and table_names to target).