Several updates to table management and updates from GCP BigQuery.
Credentials management will now take either an explicit reference to a credential.json, or a yaml file, in most cases, with the aim of resolving a valid client from a call to a service_account.Credentials object - this will eliminate authentication errors where the user if warned that the client must be instantied with credentials created in the google-auth-library.

Several TableManagement and getTableData updates. Noteable, the getData function now has support for infinity, which will get all rows, as well as n number of rows. Additionally, the function now has placeholder None arguments, which can be overridden with the location of the credentials.json, the project name, GCP dataset and the table name.

This allows the user to specify the number of rows they want, from the default file in the config-yaml.yaml. Over-riding the positional arguments allows for different BigQuery datasets and tables... potentially even different projects, if a different credential.json is provided!