Credentials management will now take either an explicit reference to a credential.json, or a yaml file, in most cases, with the aim of resolving a valid client from a call to a service_account.Credentials object - this will eliminate authentication errors where the user if warned that the client must be instantied with credentials created in the google-auth-library.
Updated DFS to exist in the correct position within metabeaver's new Search Algorithms-Trees directory structure.

Updated the getTable such that it takes an optional argument for the configuration file. This means we can dynamically create and load a configuration file, and then pass the file to getTableData in GoogleCloudPlatform.Bigquery. If set to None, metabeaver will attempt to search for a config.yaml file.

Integrated some support for Neo4J database visualisation. The explore database file will list the first 100 unique nodes and their relationships.