Added logging support!
metabeaver now supports a Logger class. This is available in OperationBeaver.LogCollector.defaultLogger.py and is a Singleton class that will persist across an application. It will log files from wherever it is called up to the highest level directory it can find, stopping when it hits a setup.py or .git file.

If run in a Docker container, it will create log files at the app.logFile location.

The class formats debug logs in the following format:

date - time - index number - python script - TYPE OF LOG (usually debug) - the print/log message.

Added functionality for file search within Docker images on the host system that match a text pattern, writing files found on the Docker container back to the host.

Added support to install pip directly, should it fail from the command line.